transwcd/requirements/fastvit_metric.txt
#torch==1.8.1
#torchvision==0.9.1
imageio==2.6.1
joblib==1.1.1
matplotlib
#mmcv==2.0.1
#mmcv_full==1.6.2
numpy==1.24.3
omegaconf==2.0.0
opencv_python==4.6.0.66
Pillow==10.0.0
#pydensecrf==1.0rc3
scikit_learn==1.3.0
tifffile==2020.9.3
timm==1.0.13
tqdm==4.46.1
## additional requirements 
tensorboard
natsort
torchmetrics
pytorch-metric-learning
wandb
pandas
